pyproject.toml
[tool.poetry]
name = "url_debugger"
version = "0.1.0"
description = ""
authors = ["Thomas Cave <thomas@cave.dev>"]
readme = "README.md"
homepage = "https://github.com/caveman280/poetry-python-template"

[tool.poetry.dependencies]
python = "^3.7.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.2"
pytest-cov = "^2.9.0"
pytest-mock = "^3.1.1"
requests-mock = "^1.8.0"

[tool.poetry.scripts]
test = "pytest:main"